LayoutTests/imported/w3c/web-platform-tests/IndexedDB/bigint_value-expected.txt

CONSOLE MESSAGE: line 41: SyntaxError: No identifiers allowed directly after numeric literal

Harness Error (FAIL), message = SyntaxError: No identifiers allowed directly after numeric literal
